fix: 系统临时邮箱配置不可删除

+++ server/internal/service/temp_mail_provider.go
@@ -90,6 +90,16 @@ func (s *TempMailProviderService) Update(id uint, name, providerType, descriptio
 }
 
 func (s *TempMailProviderService) Delete(id uint) error {
+	p, err := s.repo.FindByID(id)
+	if err != nil {
+		return err
+	}
+	if p == nil {
+		return errors.New("temp mail provider not found")
+	}
+	if p.IsSystem {
+		return errors.New("cannot delete system temp mail provider")
+	}
 	return s.repo.Delete(id)
 }
